feat: :memo: Imagenes guardaas y delimitación de usuario basico

+++ app/Config/Routes.php
@@ -17,12 +17,12 @@ $routes->group('auth', ['namespace' => 'App\Controllers'], function ($routes) {
     $routes->add('login', 'AuthController::login', ['as' => 'login']);
 
     /**
-     * Route for user registration.
+     * Route for user registration only admin's.
      *
      * Method: POST
      * Description: This route is used for user registration by providing required data.
      */
-    $routes->add('register', 'AuthController::register', ['as' => 'register']);
+    $routes->add('register', 'AuthController::register', ['as' => 'register', 'filter' => 'admin']);
 });
 
 $routes->group('users', ['namespace' => 'App\Controllers', 'filter' => 'auth'], function ($routes) {
@@ -35,26 +35,28 @@ $routes->group('users', ['namespace' => 'App\Controllers', 'filter' => 'auth'], 
     $routes->get('/', 'UsersController::index', ['as' => 'indexUser']);
 
     /**
-     * Route to retrieve a user by ID.
+     * Route to retrieve a user by ID only admin's.
      *
      * Method: GET
      * Description: This route is used to retrieve details of a user by their ID.
      *              It accepts a numeric user ID as a parameter.
      */
-    $routes->get('(:num)', 'UsersController::show/$1', ['as' => 'showUser']);
+    $routes->get('(:num)', 'UsersController::show/$1', ['as' => 'showUser', 'filter' => 'admin']);
 
     /**
-     * Route to partially update a user.
+     * Route to partially update a user only admin's.
      *
-     * Method: PATCH
+     * Method: PATCH (POST)
      * Description: This route is used to partially update the details of a user
      * by their numeric ID. It accepts a JSON or form data payload with the fields
      * to be updated.
+     * IMPORTAN: Changed method to post for method compatibility in Codeigniter4.
+     * If patch works but you cannot enter the data through form-data only because of x-www-form-urlencoded
      */
-    $routes->patch('(:num)', 'UsersController::update/$1', ['as' => 'updateUser']);
+    $routes->post('(:num)', 'UsersController::update/$1', ['as' => 'updateUser', 'filter' => 'admin']);
 
     /**
-     * Route to partially delete a user.
+     * Route to partially delete a user only admin's.
      *
      * Method: DELETE
      * Description: This route is used to partially delete a user's account
@@ -62,14 +64,25 @@ $routes->group('users', ['namespace' => 'App\Controllers', 'filter' => 'auth'], 
      * their account while keeping their data for potential recovery.
      * It accepts the user's ID as a parameter.
      */
-    $routes->delete('(:num)', 'UsersController::destroy/$1', ['as' => 'destroyUser']);
+    $routes->delete('(:num)', 'UsersController::destroy/$1', ['as' => 'destroyUser', 'filter' => 'admin']);
 
     /**
-     * Route to restore a soft-deleted user.
+     * Route to restore a soft-deleted user only admin´'s.
      *
      * Method: PATCH
      * Description: This route is used to restore a user who has been soft-deleted, providing their numeric ID.
      * Restoration allows re-enabling the user's account. It accepts the ID of the deleted user as a parameter.
      */
-    $routes->patch('restore/(:num)', 'UsersController::restore/$1', ['as' => 'restoreUser']);
+    $routes->patch('restore/(:num)', 'UsersController::restore/$1', ['as' => 'restoreUser', 'filter' => 'admin']);
+
+    /**
+     * Route for generating a PDF file containing a list of users.
+     *
+     * HTTP Method: GET
+     * Controller: PdfController
+     * Method: generateUserListPdf
+     *
+     * @see \App\Controllers\PdfController::generateUserListPdf()
+     */
+    //$routes->get('pdf', 'UsersController::generateUserListPdf');
 });
